build system: avoid build failure during bisection
Commit 4bdc914ff (build system: fix compiler warnings) added a
test on the return value of fgets() in split-include.c.

During bisection it's possible to go back to a state where a
configuration value didn't exist.  This results in an empty
include file corresponding to the missing feature.  If a
subsequent bisection returns to a state where the feature exists
split-include treats the empty file as an error and the build
fails.

Add a call to ferror() to distinguish between fgets() failing
due to an error and due to there being no data to read.

Signed-off-by: Ron Yorston <rmy@pobox.com>
Signed-off-by: Denys Vlasenko <vda.linux@googlemail.com>