Remove extra space making uptime more parser friendly - uptime.c
Removed the empty space at the beginning of the default print statement. 

From my understanding (not verified) the extra space is there so that the output aligns with some more commands like `top vmstat mpstat iostat netstat df free` and it was passed along overtime for backwards compatibility.
The problem is those utils got rid of the space and now `uptime` is not aligned with nothing.